fin sans csv

+++ src/models/Atelier/Piece/piece.model.js
@@ -6,8 +6,8 @@ const Piece = sequelize.define(
     {
         id: { primaryKey: true, type: DataTypes.INTEGER, autoIncrement: true },
         libelle: { type: DataTypes.STRING, allowNull: false },
-        prix_vente: { type: DataTypes.INTEGER, allowNull: true },
-        prix_catalogue: { type: DataTypes.INTEGER, allowNull: true },
+        prix_vente: { type: DataTypes.FLOAT, allowNull: true },
+        prix_catalogue: { type: DataTypes.FLOAT, allowNull: true },
         stock: { type: DataTypes.INTEGER, allowNull: true },
         unite: { type: DataTypes.STRING, allowNull: true },
         type: { type: DataTypes.INTEGER, allowNull: false },
