+++ src/models/Commerce/LigneCmdVente/ligne_vente.model.js
@@ -6,7 +6,7 @@ const Ligne_Vente = sequelize.define(
     {
         id: { primaryKey: true, type: DataTypes.INTEGER, autoIncrement: true },
         libelle: { type: DataTypes.STRING, allowNull: false },
-        prix_vente: { type: DataTypes.INTEGER, allowNull: false },
+        prix_vente: { type: DataTypes.FLOAT, allowNull: false },
         quantite: { type: DataTypes.INTEGER, allowNull: false },
         unite: { type: DataTypes.STRING, allowNull: false },
     },
